Controller boilerplate

+++ src/main/java/com/example/repository/AccountRepository.java
@@ -3,8 +3,5 @@ package com.example.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.entity.Account;
 
-public interface AccountRepository extends JpaRepository<Account, Integer> {
-
-    
-
+public interface AccountRepository extends JpaRepository<Account, Long> {
 }
